Refactor font stack configuration type.
Previously font stacks were taken as a string (e.g. "'Helvetica Neue', Arial, sans-serif") and unquoted by the typography helper when used. This was originally done because the unescaped use of commas within a map is a syntax error, but it created issues as it required the value to ALWAYS be quoted, even if specified in a context where it wasn't a requirement, such as a variable.

Anyway, it turns out that quoting was completely unnecessary as Sass supports passing in lists for this purpose (e.g. ("Helvetica Neue", Arial, sans-serif)), so now we do that.